refactor: turn Sequence into a struct
Change the Sequence type from an interface into a struct.

This will allow for easier addition of useful methods, such as
materialize, and the addition of data / options that can speed up some
operations.

Basically, made the named change, but then needed to:
 - Rework Volatile
   - Use above in lines and chans
 - Move around code between files
 - Fix tests
 - Updated documentation
