+++ sequence.go
@@ -2,8 +2,6 @@ package sequence
 
 import (
 	"errors"
-
-	"github.com/cookieo9/sequence/tools"
 )
 
 var (
@@ -15,60 +13,22 @@ var (
 // A Sequence represents a functionally immutable series of values. The
 // sequence can only be used to fetch the values stored within, but the
 // sequence itself can't be used to make changes.
-type Sequence[T any] interface {
-	Each(func(T) error) error
+type Sequence[T any] struct {
+	source   func(func(T) error) error
+	volatile bool
 }
 
-// Each returns a function to iterate over the sequence. The callback to
-// the produced function is for users to receive and handle each value from
-// the sequence, and return an error if processing should stop. The first
-// error produced by the sequence, either an error from the callback or
-// an error producing values will be returned. The [ErrStopIteration] error
-// will not be returned from the iteration function.
-func Each[T any](s Sequence[T]) func(func(T) error) error {
-	return func(f func(T) error) error {
-		if err := s.Each(f); !errors.Is(err, ErrStopIteration) {
-			return err
-		}
-		return nil
-	}
-}
+// IsVolatile returns true if the sequence is volatile, meaning that it should
+// only be iterated over once. It is an error to call Each/EachSimple/Iterator
+// more than once. Volatile sequences can be created with the [Volatlie]
+// function.
+func (s Sequence[T]) IsVolatile() bool { return s.volatile }
 
-// EachSimple is like [Each], where a function is returned to iterate over
-// the contents of the sequence, but the callback is in a simpler true/false
-// form. An error generating a value will be returned, and iteration stopped.
-// When the callback returns false, an ErrStopIteration error is passed
-// through the sequence to anyone who cares about errors, but will be dropped
-// before returning from EachSimple as it doesn't represent a real error.
-func EachSimple[T any](s Sequence[T]) func(func(T) bool) error {
-	fn := Each(s)
-	return func(f func(T) bool) error {
-		err := fn(tools.Compose(f, b2Err))
-		if !errors.Is(err, ErrStopIteration) {
-			return err
-		}
-		return nil
-	}
-}
-
-// Iterator is a wrapper on [EachSimple] and is used to work with the upcoming
-// language extension that allow functions in a for-range statement. Thus, when
-// avaliable, one may write "for x := range Iterator(seq) {}" to use a sequence
-// directly in a for-loop.
-//
-// Errors when generating the contents of the sequence will result in a panic.
-func Iterator[T any](s Sequence[T]) func(func(T) bool) {
-	fn := EachSimple(s)
-	return func(f func(T) bool) {
-		if err := fn(f); err != nil {
-			panic(err)
-		}
-	}
-}
-
-func b2Err(b bool) error {
-	if b {
-		return nil
-	}
-	return ErrStopIteration
+// Each iterates over every item in the sequence calling the passed callback
+// with each item. An error returned from the callback, or one that arose from
+// the processing of the sequence will be returned if they arise and iteration
+// will stop. Unlike the top-level functions that iterate over sequences, this
+// method will not on it's own handle [ErrStopIteration] and return it instead.
+func (s Sequence[T]) Each(f func(T) error) error {
+	return s.source(f)
 }
